I had some name collisions, and that was crashing it.
Oops. I renamed some things, but some other thing with that name was overriding it and screwing
everything up. I fixed it, then realized that the thing that was breaking everything wasn't
even necessary, so I removed it altogether. Now it works, and it's cleaner. So I guess it's a win
overall.